♻️refactor(k8s): remove circular import
